Moved code into more defined directories.
