+++ run.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+
+(cd vanilla-frontend && npx tsc -w) & (cd backend/forum && go run .) & (cd backend/chat && go run .)
