+++ vanilla-frontend/src/main.ts
@@ -1,4 +1,4 @@
-import { Auth } from "./components/auth.js"
+import { Auth } from "./components/authorization/auth.js"
 
 export const backendUrl = "http://localhost:8080"
 export const superDivision = document.getElementById(
